tests: Use home directory
This ensures the target is existing directory and that it is on
a file system, which is native to the underlying operation system.
E.g. when trying to read a directory as a file on NFS on Windows
(can happen easily in VirtualBox, for instance), the operation
won't throw an exception, but will read an empty string.